Initalize struct ctype argument to zero in to_cdata.
Some callers of to_cdata check the pointers field of ctype which caused
undefined behvaior. For example, multiplying a number by a ctype<long long>
would sometimes give a "can't operate on a pointer value" error because
the pointers field might be 1 in the case that the argument wasn't a ctype.